[MM-48862] Unify style and position of arrows for month selection on date picker
