+++ sass/components/_day-picker.scss
@@ -13,6 +13,9 @@
             border-bottom: 1px solid transparent;
             margin: 0 0.5rem 0.7rem;
             text-align: center;
+            color: var(--center-channel-color);
+            font-size: 14px;
+            font-weight: 600;
 
             > div {
                 font-size: inherit;
